Python program to scan computer and compare to list of threats
This program scans computer's .exe files and compiles them in a list. Then it compares that list with the txt extracted from the csv database through fullread.py. Finnaly, the program prints a txt with the .exe matches of possible threats, so that they can be further analyzed by the user.